pyproject.toml
[tool.poetry]
name = "racecar"
version = "0.1.0"
description = ""
authors = ["Cameron Reid <csreid.cr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^4.2.2"
torch = "^2.3.1"
numpy = "<2"
swig = "^4.2.1"
ray = {extras = ["rllib"], version = "^2.30.0"}
autorom = {extras = ["accept-rom-license"], version = "^0.6.1"}
tqdm = "^4.66.4"
stable-baselines3 = {extras = ["extra"], version = "^2.3.2"}
gymnasium = {version = "0.28.1", extras = ["box2d"]}
pyzmq = "^26.2.0"
poethepoet = "^0.27.0"
ruff = "^0.6.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
torchvision = "~0.18"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
